reafactor(service): standardize service naming
