+++ pet-clinic-data/src/main/java/com/hajres/petclinic/service/map/OwnerMapService.java
@@ -12,11 +12,11 @@ import java.util.Set;
 
 @Service
 @Profile("default")
-public class OwnerServiceMap extends AbstractMapService<Owner, Long> implements OwnerService {
+public class OwnerMapService extends AbstractMapService<Owner, Long> implements OwnerService {
     private final PetTypeService petTypeService;
     private final PetService petService;
 
-    public OwnerServiceMap(PetTypeService petTypeService, PetService petService) {
+    public OwnerMapService(PetTypeService petTypeService, PetService petService) {
         this.petTypeService = petTypeService;
         this.petService = petService;
     }
