+++ pet-clinic-data/src/main/java/com/hajres/petclinic/service/map/PetMapService.java
@@ -1,7 +1,7 @@
 package com.hajres.petclinic.service.map;
 
-import com.hajres.petclinic.model.PetType;
-import com.hajres.petclinic.service.PetTypeService;
+import com.hajres.petclinic.model.Pet;
+import com.hajres.petclinic.service.PetService;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
@@ -9,9 +9,9 @@ import java.util.Set;
 
 @Service
 @Profile("default")
-public class PetTypeServiceMap extends AbstractMapService<PetType, Long> implements PetTypeService {
+public class PetMapService extends AbstractMapService<Pet, Long> implements PetService {
     @Override
-    public Set<PetType> findAll() {
+    public Set<Pet> findAll() {
         return super.findAll();
     }
 
@@ -21,17 +21,17 @@ public class PetTypeServiceMap extends AbstractMapService<PetType, Long> impleme
     }
 
     @Override
-    public void delete(PetType object) {
+    public void delete(Pet object) {
         super.delete(object);
     }
 
     @Override
-    public PetType save(PetType object) {
+    public Pet save(Pet object) {
         return super.save(object);
     }
 
     @Override
-    public PetType findById(Long id) {
+    public Pet findById(Long id) {
         return super.findById(id);
     }
 }
