+++ pet-clinic-data/src/main/java/com/hajres/petclinic/service/map/SpecialityMapService.java
@@ -9,7 +9,7 @@ import java.util.Set;
 
 @Service
 @Profile("default")
-public class SpecialityServiceMap extends AbstractMapService<Speciality, Long> implements SpecialityService {
+public class SpecialityMapService extends AbstractMapService<Speciality, Long> implements SpecialityService {
     @Override
     public Set<Speciality> findAll() {
         return super.findAll();
