+++ pet-clinic-data/src/main/java/com/hajres/petclinic/service/map/VetMapService.java
@@ -11,10 +11,10 @@ import java.util.Set;
 
 @Service
 @Profile("default")
-public class VetServiceMap extends AbstractMapService<Vet, Long> implements VetService {
+public class VetMapService extends AbstractMapService<Vet, Long> implements VetService {
     private final SpecialityService specialityService;
 
-    public VetServiceMap(SpecialityService specialityService) {
+    public VetMapService(SpecialityService specialityService) {
         this.specialityService = specialityService;
     }
